Add notification model and controller; update comment and like models to include content type; refactor subscription routes and controller methods

+++ src/controllers/comment.controller.js
@@ -2,21 +2,20 @@ import { ApiError } from "../utils/apiError.js";
 import asyncHandler from "../utils/asyncHandler.js";
 import { Comment } from "../models/comment.model.js";
 import { ApiResponse } from "../utils/apiResponse.js";
-import mongoose from "mongoose";
 
 export const postComment = asyncHandler(async (req, res) => {
-  const { text, content = "video", contentId } = req.body;
+  const { text, contentType = "Video", contentId } = req.body;
 
   if (!text || !contentId) {
     throw new ApiError(400, "Comment Text And Content Required");
   }
 
-  const capitalizeContent = content.charAt(0).toUpperCase() + content.slice(1);
+  const capitalizeContent = contentType.charAt(0).toUpperCase() + contentType.slice(1);
 
   const comment = await Comment.create({
     text,
     commentedBy: req.user._id,
-    content: capitalizeContent,
+    contentType: capitalizeContent,
     contentId,
   });
 
@@ -31,7 +30,7 @@ export const postComment = asyncHandler(async (req, res) => {
 
 export const editComment = asyncHandler(async (req, res) => {
   const { text } = req.body;
-  const commentId = req.params.id;
+  const commentId = req.params.commentId;
 
   if (!commentId) {
     throw new ApiError(400, "Comment ID Is Required.");
@@ -39,6 +38,12 @@ export const editComment = asyncHandler(async (req, res) => {
 
   if (!text) {
     throw new ApiError(400, "Comment Text Is Required.");
+  }
+
+  const commentOwner = await Comment.findById(commentId);
+
+  if (!commentOwner || commentOwner.commentedBy.toString() !== req.user._id.toString()) {
+    throw new ApiError(403, "Unauthorized to Edit This Comment.");
   }
 
   const comment = await Comment.findByIdAndUpdate(
@@ -57,10 +62,16 @@ export const editComment = asyncHandler(async (req, res) => {
 });
 
 export const deleteComment = asyncHandler(async (req, res) => {
-  const commentId = req.params.id;
+  const commentId = req.params?.commentId;
 
   if (!commentId) {
     throw new ApiError(400, "Comment ID Is Required.");
+  }
+
+  const commentOwner = await Comment.findById(commentId);
+
+  if (!commentOwner || commentOwner?.commentedBy.toString() !== req.user._id.toString()) {
+    throw new ApiError(403, "Unauthorized to Delete This Comment.");
   }
 
   const comment = await Comment.findByIdAndDelete(commentId);
@@ -75,18 +86,25 @@ export const deleteComment = asyncHandler(async (req, res) => {
 });
 
 export const getCommentsByContent = asyncHandler(async (req, res) => {
-  const { contentId, content = "video" } = req.params;
+  const { contentId, contentType = "video" } = req.params;
 
   if (!contentId) {
     throw new ApiError(400, "Content ID Is Required.");
   }
 
-  const capitalizeContent = content.charAt(0).toUpperCase() + content.slice(1);
+  const capitalizeContent = contentType.charAt(0).toUpperCase() + contentType.slice(1);
 
   const comments = await Comment.find({
     contentId,
-    content: capitalizeContent,
-  }).populate("commentedBy", "name profilePicture");
+    contentType: capitalizeContent,
+  })
+    .populate("commentedBy", "username avatar")
+    .limit(20);
+
+  const totalCount = await Comment.countDocuments({
+    contentId,
+    contentType: capitalizeContent,
+  });
 
   if (!comments || comments.length === 0) {
     throw new ApiError(404, "No Comments Found.");
@@ -94,31 +112,30 @@ export const getCommentsByContent = asyncHandler(async (req, res) => {
 
   res
     .status(200)
-    .json(new ApiResponse(200, comments, "Comments Retrieved Successfully."));
+    .json(new ApiResponse(200, { comments, totalCount }, "Comments Retrieved Successfully."));
 });
 
 export const getCommentOfUser = asyncHandler(async (req, res) => {
+  const { contentType = "Video" } = req.query;
   const userId = req.user._id;
 
   if (!userId) {
     throw new ApiError(400, "User ID Is Required.");
   }
 
-  const comments = await Comment.aggregate([
-    {
-      $match: {
-        commentedBy: mongoose.Schema.Types.ObjectId(userId),
-      },
-    },
-    {
-      $lookup: {
-        from: "users",
-        localField: "commentedBy",
-        foreignField: "_id",
-        as: "userDetails",
-      },
-    },
-  ]);
+  if (!contentType) {
+    throw new ApiError(400, "Content Type Is Required.")
+  }
+
+  const capitalizeContent = contentType.charAt(0).toUpperCase() + contentType.slice(1);
+
+  const comments = await Comment.find({ commentedBy: userId, contentType: capitalizeContent })
+    .populate("commentedBy", "username avatar");
+
+  const totalCount = await Comment.countDocuments({
+    commentedBy: userId,
+    contentType: capitalizeContent
+  })
 
   if (!comments || comments.length === 0) {
     throw new ApiError(404, "No Comments Found For This User.");
@@ -127,6 +144,6 @@ export const getCommentOfUser = asyncHandler(async (req, res) => {
   res
     .status(200)
     .json(
-      new ApiResponse(200, comments, "User Comments Retrieved Successfully.")
+      new ApiResponse(200, { comments, totalCount }, "User Comments Retrieved Successfully.")
     );
 });
